requirements.txt
requests==2.32.3
ipykernel
python-dotenv
openai==1.59.6
pydantic==2.10.5
pydantic-ai==0.0.18
pydantic-ai-slim==0.0.18
pydantic_core==2.27.2
supabase==2.11.0
Crawl4AI==0.4.247
streamlit==1.41.1
logfire==3.1.0
logfire-api==3.1.0
httpcore==1.0.7
httpx==0.27.2
crewai==0.28
crewai_tools==0.1.6
langchain_community==0.0.29
langchain-openai
helper
setuptools
langgraph
langchain-core
trustcall